[SH] Add a new command "copybbt" (for "skipping" bad blocks in NAND booting)
This update is to add a new command "copybbt", which is useful in
"patching" a u-boot.bin in RAM, before burning it to NAND flash. The
ultimate requirement is to copy a sub-set of the master Bad Block Table
(BBT) into a "skip_bb" structure located in physical block #0 of NAND.

There are two sub-commands added:

	U-Boot> copybbt info <addr>
and	U-Boot> copybbt copy <addr>

Both commands take a mandatory parameter which is the address in RAM where
the u-boot.bin binary was loaded. Typically, this parameter will just be
"$load_addr".  [The "skip_bb" structure will typically be located at an
offset of 0x00C0 bytes after the passed address.]

The "info" sub-command will (check for correctness and then), dump out
the contents of the "skip_bb" structure, in a human friendly format.

The "copy" sub-command will (check for correctness and then) copy a
sub-set of the master BBT into the array in the "skip_bb" structure.
The modified copy may then be burned into NAND flash.

For more details, please see "cpu/sh/cmd_copy_bbt.c".

As an example of usage, the following may help. Note the commands marked
with "***" are not needed, but should assist with initial comprehension.

	U-Boot> nfs $load_addr /export/u-boot.bin
***	U-Boot> copybbt info $load_addr
***	U-Boot> nand bad
	U-Boot> copybbt copy $load_addr
***	U-Boot> copybbt info $load_addr
	U-Boot> nand write.jffs2 $load_addr 0 $monitor_len
***	U-Boot> nand dump 0

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>