[SH] Finalized support for MXIC MX25Lxxxx SPI serial flash.
This update tidies up the previous commit which added initial support
for the Macronix MX25L12855E SPI serial flash. The code is slightly
better structured, and now supports a few (but not all) members of the
MXIC MX25Lxxxx family of Macronix's secure SPI serial flash devices.

This update includes support for boot-from-MX25L12855E on both 29-bit
and 32-bit modes, which has been tested on a STx5197-CAB-V1.2A-065 board.

NOTE: the SPI Boot Mode controller configuration for the MXIC serial
flash is the same as that used for ST serial flash. This appears to work
correctly, but it is not know if this is an optimal configuration for
the MXIC devices or not. If problems subsequently manifest themselves
for the MXIC parts, then this configuration should be verified.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>