[SH] Remove vestigial FPU instructions from start.S for ST40.
As U-boot no longer uses the FPU (at all), then there was a minor
vestigial inconsistency, that "start.S" was still initializing the FPSCR,
but the code was still compiled "-m4-nofpu".

With an updated compiler which enforces the "--isa" in the assembler,
the compilation will fail.  This update simply removes the FPU-specific
instructions from "start.S", allowing the code to be build without an FPU.

Currently, the code is compiled with:

	-m4-nofpu -U__sh3__ -D__SH4__=1

However, with the next compiler, it should be possible to use:

	-m4 -mno-implicit-fp

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>