[SH] Added support for SPI serial flash on the PDK-7105 board.
Added support for SPI serial flash on the PDK-7105 board, which uses
STMicroelectronics' STx7105 SoC device.

This update exclusively uses "bit-banging" (with PIOs) and does not use
the SSC to talk to the SPI serial flash device (c.f. the STx5197).

This has only been tested on a single STi7105SDK-V1.2-022 board. Some
of the code has been re-factored, and really needs to be properly tested
for potential regressions on the STx5197 boards (MB704 & 5197CAB).

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>