[bugfix] Increase CFG_ENV_SIZE for NAND flash for the PDK-7105 board.
Fixed a bug in "board/pdk7105/pdk7105.c", whereby CFG_ENV_SIZE was too
small (less than one large-page block-size). The result of this bug was
that the "saveenv" command failed, but other writes to NAND succeed.

This update just increases the value of CFG_ENV_SIZE if it is less than
the size of one (large-page) NAND block (i.e. 128 KiB).

Also, for both NAND and NOR flashes, the type of the flash device is
now also printed when "saveenv" is used. This just helps the user to
know he is writing the correct flash type (NOR, NAND or SPI).

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>