[SH] Completed support for boot-from-SPI on the PDK-7105 board.
This update completes the "boot-from-SPI" serial flash on
STMicroelectronics' PDK-7105 Board, with the STx7105 SoC.

Previously, booting did not work, this is the first version which is
demonstrably bootable (from SPI serial flash, in 29-bit mode).

This has only been tested on a single STi7105SDK-V1.2-022 board, and
only in 29-bit mode.  More testing is required, and re-testing for the
STx5197 is also needed.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>