[SH] Added support for STMicroelectronics' IPIDTV7105 board.
This update adds support for the IPIDTV7105 board which uses
STMicroelectronics' STX7105 SoC device.

This board supports both NAND and SPI flash (but not NOR flash).
Reading and writing to both NAND and SPI has been tested, as well as
testing of USB, SATA and Ethernet. Testing was done in both 29-bit and
32-bit modes, both downloaded via JTAG, and booted from NAND flash.

A minor board modification is required to change the (default)
boot-from-NAND behaviour to boot-from-SPI. As I only have access to boards
that have not been modified, then only boot-from-NAND has been completed &
tested. Currently, this updates does not (even) build for boot-from-SPI,
and a few lines in include/configs/ipidtv7105.h will need to be modified
once a boot-from-SPI board is made available to test.

TODO: boot-from-SPI serial flash.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>