[FIX] Avoid a nasty stack corruption if no NOR flash device found.
We need to ensure that sizeof(cword) >= info->portwidth, otherwise,
flash_make_cmd() will scribble over memory it should not!  This results
in a stack corruption, and madness follows...

This only occurs if probing for NOR flash failed to find any such devices,
and portwidth ends up being 16. However, we then try and issue a "reset"
command, which copies 16-bytes into a 8-byte pointer union on the stack,
and BANG! Instead, we just issue an error, and do not issue the command.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>