[SH] Redefine STPIO_GET_PIN() to be consistent with the other PIO macros.
The parametrizable macros to get/set PIOs were inconsistent.
Specifically, most take as their first parameter, the base address of the
specific PIO controller to use.  However, in the case of STPIO_GET_PIN(),
the first parameter instead was the numerical port number.

This patch simply redefines the API for STPIO_GET_PIN(), to make it
consistent with all the other PIO macros, so they all behave similarly.

For further information please see:

	https://bugzilla.stlinux.com/show_bug.cgi?id=6864

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>