[SH] Fix standalone application (bugzilla #6208)
This patch fixes a bug in the address calculation for the EXPORT_FUNC
macro, for the SH arch, as used by U-boot's standalone applications.

Without this fix, the newer aggressive compiler optimizer, optimized
the code in such a way as some of the previous assumptions no longer held.

For further details, please see:

	https://bugzilla.stlinux.com/show_bug.cgi?id=6208

Signed-off-by: Angelo Castello <angelo.castello@st.com>