[SH] Now use CFG_SDRAM_BASE to build default standalone applications.
This update builds on the previous patch for "standalone applications".
There are two main aspects, of this patch.

Firstly, the assembler code in the macro EXPORT_FUNC (examples/stubs.c)
has been tidied up a bit, and the comments improved. Both of these should
make the code more transparent, and more maintainable!

Secondly, the main update is the automatic determination of choosing
a sensible value for the makefile macro "LOAD_ADDR", which is used as
the address to link the standalone application. The problem with this,
is that the default standalone application "hello_world", needs to be
built at one address.  However, that address can very from board to board,
and also vary depending if the system is in 29-bit or 32-bit mode. Hence,
it was previously necessary to modify the examples/Makefile to change it.

The U-boot infrastructure did not provide a simple means of passing any
valid address from the board's configuration file into the "examples"
tree.  The infrastructure to do this has now been added.

For SH (and ST40) devices, the default link address will now be the same
as the configuration option CFG_SDRAM_BASE, which is usually the same as
U-boot's environment variable "load_addr". Hence, the following should
now work for all ST40 based systems.

	host% make distclean
	host% make <xxx>_config
	host% make
	host% cp -p examples/hello_world.bin /export
	host% sh4-linux-gdb -ex "sh4tp <xxx>" u-boot

	U-boot> nfs $load-addr /export/hello_world.bin
	U-boot> go $load_addr arg1 arg2 arg3 arg4

This has been tested in both 29-bit and 32-bit modes, on one board.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>