[SH] Added better checking for "bootm" (address range validation)
	Modified "bootm" for the ST40, to verify that when a gzip'ed
kernel is being uncompressed, that that the memory range being used
does not "overlap" with either U-boot or the compressed ".ub" file,
and that it is sensible (i.e. in RAM).

	The following tests have (logically) been added:

		assert (Uload >= CFG_SDRAM_BASE);
		assert (Uend < CFG_MEMTEST_END);
		assert ( (Cload > Uend) || (Uload > Cend) );

	If any of these tests fail, then the kernel will not be
uncompressed, the boot will not occur, and a suitable diagnostic will be
issued. Hopefully, users will be able to correct this class of problem
a lot earlier than before.

	As a result of this, the "CFG_BOOTM_LEN" is no longer used in
trying to determine if the image fits into memory. The aforementioned
checks are more sophisticated, than the simple "CFG_BOOTM_LEN" ones.

	For more details, then please have a look at:

	https://bugzilla.stlinux.com/show_bug.cgi?id=6280

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>