zlib: add watchdog reset call
This patch adds watchdog reset call to allow its invocation during
decompression phase. This control was present on old zlib version and
here it is back-ported for those relevant routines.

This patch is sent as a separate zlib one because it was not tested due
to lack of a specific board.  zlib patches will be unified just in one
when this will be validated through tests.

Patch back-ported (by Sean McGoogan <Sean.McGoogan@st.com>) from:
http://git.denx.de/?p=u-boot.git;a=commit;h=253cb831f5861358a7fa673305cdf7ded1096f44

Signed-off-by: Giuseppe Condorelli <giuseppe.condorelli@st.com>
Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>