[SH] Additional speed gain by compiling zlib.c with "-O2".
Empirical measurement shows that just by compiling this one file for
speed (instead of space), gives ~5% improvement in boot time, but at a
cost only of about 3 KiB in code size. If size is critical, then this
patch may be undone, and "-Os" used instead.

This patch is specific to STMicroelectronics, and gives an additional
speed gain just by compiling the one file lib_generic/zlib.c with "-O2",
instead of the default "-Os" optimization flag.

*NOTE* this patch should not be blindly merged into the master git
repository, as it changes the compile options for *all* architectures, and
it has only been tested specifically with ST40 SoCs.  This patch probably
will need to be re-written to be acceptable to the top-level maintainers.

Signed-off-by: Giuseppe Condorelli <giuseppe.condorelli@st.com>
Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>