[SH] add support for a ndelay() (nano-second delay) function for SPI
Currently, the S/W "bit-banging" SPI driver runs at a maximum speed of
500 kHz, as it has to issue a udelay(1) after each time the SPI clock
is toggled (taking 2 invocations per clock cycle).

In an attempt to speed up the bit-banging SPI driver, a sub micro-second
delay capability is required. The SH4 has a TMU which can divide the
peripheral clock down by 5 different factors: 4, 16, 64, 256, 1024.
Prior to this patch, each configuration file has a line like:

	#define TMU_CLK_DIVIDER		64

This divided the Peripheral clock by 64, so a 100 MHz peripheral clock
would result in the TMU being clocked at 1.5625 MHz, giving a TMU "tick"
of 0.64us, and an overflow after ~46 minutes.

However, to increase the precision, then it is required to increase the
TMU clock, by reducing the clock divider. Using a clock divider of 4,
then we have a TMU clock of 25 MHz (still assuming a peripheral clock
of 100MHz), a TMU "tick" of 40ns, and a overflow of ~171 seconds. A 40ns
delay, would allow for a maximum of 12.5 MHz bit-banging SPI clock. Which
is acceptable.

This patch, provides a new function ndelay(), which will cause a delay
specified in nano-seconds (rounded to the next highest multiple of 40ns,
with a 100MHz PClock). The value of TMU_CLK_DIVIDER now defaults to 4 in
lib_sh/time.c, although, users may over-write this is the configuration
files, if required.

Although this patch will reduce the maximum delay and timer to ~171
seconds, it is not believed that this causes any real problems in U-boot!
Obviously, faster peripheral clocks will reduce the maximum delay +
timer values proportionally.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>