[SH] Added "boot-from-SPI" for ST's IPIDTV7105 board.
This update adds tested support for the "boot-from-SPI" for
STMicroelectronics' IPIDTV7105 board.

It is required to modify the board (soldering) to convert a IPIDTV7105
board between boot-from-NAND, and boot-from-SPI. Now that I have access
to both flavours of boards, I was able to finish the port, and empirically
verify that both modes are functional.

Testing took place on both boot types, in both 29-bit and 32-bit modes. In
particular, it is now possible to boot-from-SPI with the environment in
SPI in 32-bit mode (this required an additional PMB slot to be defined).

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>