[SH] Optimized NAND read using FLEX-mode with cached 32-byte access
This optimization has been shown (on a IPIDTV7105) to reduce the time
taken to read 16MiB from NAND by 42% (almost doubling performance).

In FLEX-mode, we can either read from the Flex-mode data register
(ST40_EMI_NAND_FLEX_DATA), over the STBus with either a 4-byte read
(LD4) or a 32-byte read (LD32) bus opcode.  Using the LD32 bus opcode
amortises the cost of the bus latency over 32-bytes, instead of only 4
bytes, so one can achieve a higher throughput, using the LD32 bus opcode.

However, in order to realize this potential improvement, one needs
to access the ST40_EMI_NAND_FLEX_DATA via the CPU's operand (data)
caches. This in turn requires a TLB (or PMB) address translation to
be configured and enabled.  This update, creates a single (read-only)
1KiB TLB mapping, including mapping the ST40_EMI_NAND_FLEX_DATA register
into a P3 (cachable + translatable) virtual address, so we can utilize
the LD32 opcode, by mapping ST40_EMI_NAND to 0xC0000000.

It should be noted that the aforementioned TLB mapping is set up only
once, during the NAND initialization. However, the TLB is only enabled
(MMUCR.AT) when were are actively reading from NAND. That is the TLB is
only used when we are in stm_flex_read_buf(). So, when U-boot is in the
"shell", then it will appear that the TLB is turned off.

We could use a similar optimization for writing, however, only read
performance is more important for booting. So, for the time being,
this optimization only applies to reading from NAND.

Also, this optimization is only enabled if the amount of data to be
read from the NAND is a multiple of one cache line (32 bytes). Most
transactions have this property, so most buffered reads will benefit
from this transaction.

In order to enable this performance gain, the configuration macro
CONFIG_SH_NAND_USES_CACHED_READS needs to be defined.

This has only been tested on one board (IPIDTV7105), but in both 29-
and 32-bit modes.

Please note, an *experimental* version of the code to copy from the
cache-line to RAM is also included, but commented out with "#if 0". This
seems to give only a very small further improvement, but further testing
and validation would be required. In particular, this experimental code
uses the FPU registers to effect the copy, but no attempt is made to
preserve their contents, or ensure the FPU is present & enabled.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>