[SH] Added support for the FLI7510 Development Board
This update provide a "first pass" update for STMicroelectronics' FLI7510
(Freeman Lite) Development Board (FLDB_GPD201).

The following features have been added and successfully tested:
	Serial (UART),
	Ethernet (GMAC + PHY),
	NOR Flash,
	NAND Flash (both FLEX-mode & bit-banging),
	32-bit (SE) mode.

The following is not yet working: USB & SPI. The code for USB has been
written, but (thus far) it is still not working!

There are some outstanding issues with the provided target-packs
for this board.  For instance, the first two EMI banks should both
(presumably) be 32MiB, however all EMI banks are set to be 16 MiB. The
provided target-packs are missing key information for the EMI timings
configurations. Also it appears that the emi.EMI_BANK<x>_BASEADDRESS.poke
implementations are wrong, and write to the wrong addresses, which can
result in NAND failing to work! Also, due to some DDR issue, the PMB
configuration need to use the LMI @ 0x40000000.

As a result, no attempt has been made to boot from FLASH, so further
work+testing is required, once an updated target-pack is available.
Similarly, only 32-bit (SE) mode has been attempted to date.

Note: to get EMI working, a H/W modification to the board is required,
as well as a specific modification to fldb_gpd201.py is also required.

All the above was tested on a single FLDB_GPD201 board, with a serial
number of SZ015.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>