[SH] Fix bug with environment for the STx5197 SPI serial FLASH
It looks like when the SPI code for the STx7105 was added to U-boot for
the stm23_0044 release, that the SPI environment was inadvertently broken
for the STx5197 (and I believe only the STx5197) SoC.

I now provide this patch, which should solve this for all SoCs. This
has been tested on a STx5197 and a STX7105 board, and seems to work.

See: https://bugzilla.stlinux.com/show_bug.cgi?id=7374

This update will simply *always* use the SPI driver to read the
environment to relocate it. Previously it would try and use the SPI-Boot
mode controller to read it (with memcpy()), if appropriate. However, if
the SPI boot-mode controller has been disabled, then this copy fails. This
update potentially will slow down booting slightly, if the environment
is in SPI, but it is much safer, and should work on *all* SoCs.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>