[bug-fix] Fix problems with JFFS2 with integral "bad blocks"
Without this update, the JFFS2 code would try and read from a bad
block, which would fail. This would generate a suitable diagnostic, and
(to cut a long story short) eventually return a NULL pointer. However,
none of the code checked if the pointer was bad, and in numerous places
the code just assumed it was a legal pointer, pointing to the recently
read data.  This casual approach with NULL pointers resulted in many
different issues, and also explains how it was possible for this bug
also to "hang" the system.

The provided update fixes this problem, by replacing all such failed
reads with an equivalent length of 0xFF bytes, making the block instead
appear as "good" and "freshly erased". Due to the way JFFS2 works,
this exploitation seems to work with no problems. There should be
no more NULL pointers being de-referenced (at least for this case).
In addition, the fix will also "hide" the previous diagnostics, as
the failure mechanism is now taking appropriate action to fix-up the
underlying issue transparently.

See https://bugzilla.stlinux.com/show_bug.cgi?id=6516

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>