[SH] Added SPI support for the FLI7510 Development Board
This update provides update for STMicroelectronics' FLI7510 (Freeman Lite)
Development Board (FLDB_GPD201), by adding support for SPI serial flash.
The SPI serial flash is accessed via S/W bit-banging (not via the SSC).

In addition, this update adds support for Numonyx's M25PXxx family of SPI
serial flash devices. The capability of the smaller (4KiB) sub-sector
erase has been utilised in this update, but not support for the "dual
I/O" feature.

USB (OHCI support) is still not working!

The above was tested on a single FLDB_GPD201 board, with a serial number
of SZ015.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>