[SH] Removed support for 29-bit mode for the FLI7510 Development Board
This update removes the original support for 29-bit (non-SE) mode
in STMicroelectronics' FLI7510 (Freeman Lite) Development Board
(FLDB_GPD201). The 29-bit mode was not functional anyway, and the board
is intended to be only used in 32-bit (SE) mode anyway.

This simply removes the possible confusion that would arise if someone
built U-Boot for 29-bit mode for this board, and could not get it to work!

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>