[SH] Add support for the H/W SCC for SPI on the PDK-7105
This patch adds support for SPI over the SSC on the PDK-7105 board,
which uses STMicroelectronics' STx7105 SoC.

This provides a performance improvement of over one order of magnitude.
Initial testing suggests that using a clock frequency of 5 MHz is stable.
Corruption was observed with higher frequencies. As all testing was just
done on one PDK-7105-V1.2-002 board, then this may be unrepresentative.

Note, that by default the (slower) S/W "bit-banging" driver is enabled
by default.  To use the (faster) H/W SSC instead, then the configuration
file should be modified to undefine the macro "CONFIG_SOFT_SPI". In
addition, it may be required to modify the macro CFG_STM_SPI_FREQUENCY.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>