[ide] Spin-up ATA disks if power-up into standby feature is enabled.
This update will detect if a ATA disk has the "Power-Up in Standby"
feature set enabled, and the disk is currently spun-down, and requires
an explicit feature code (0x07) sub-command to spin-up the device.

The sub-command to spin up the disk will be automatically executed when
the "ide reset" command detects that the selected disk is spun-down.

This has been tested using a Hitachi HCC543216L9SA00 HDD, which has the
"Power-Up in Standby" feature set enabled, and required this fix in
order to access the disk. Tested on the STi7105SDK-V1.2-022 board.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>