[SH] Allow boot-from-SPI on the FLI7510 Development Board (FLDB)
This update provides working support for STMicroelectronics' FLI7510
(Freeman Lite) Development Board (FLDB_GPD201), by slowing down the SPI
clock further, to allow boot-from-SPI to boot successfully.

The SPI controller has a register which controls how much to divide
the clock by. For most SoCs + boards, setting this register to 4 is
sufficient.  However, for the FLI7510 Development Board (FLDB), the
divider register needs to be increased to a value of 6.

This update will convert:

	*ST40_EMI_SPI_CLOCKDIV = 4

to

	*ST40_EMI_SPI_CLOCKDIV = 6

for the FLI7510.

All testing was just done on one board FLDB_GPD201 (S/N: SZ015).

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>