[SH,SPI] Preparatory work for adding support for the FSM SPI Controller.
This update logically should not modify the run-time behaviour of
this code at all! This code is purely to prepare the code to support
STMicroelectronics' new FSM (Fast Sequence Mode) SPI Serial Flash
controller.

Prior to this patch, one selected between using the S/W PIO bit-banging
for SPI, or the H/W SSC controller for SPI, using just one macro:
CONFIG_SOFT_SPI. The presence of it indicated S/W bit-banging, and the
absence of it indicated the H/W SSC.

There are now 3 separate macros, and *exactly* ONE of them must be
defined, if CONFIG_SPI is enabled for the ST40. The 3 macros are:

	CONFIG_SOFT_SPI		- S/W Bit-Banging
	CONFIG_STM_SSC_SPI	- H/W using STM's SSC
	CONFIG_STM_FSM_SPI	- H/W using STM's FSM SPI Controller

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>