[SH] Correct accuracy of timers with the STx5197.
Note that there are two distinct clocks we need to deal with: The "comms"
clock and the "peripheral" clock.

	The "comms" clock is used by: ASC, SSC
	The "peripheral" clock is used by: TMU, SCIF, Watchdog

Unfortunately, due to various historical circumstances and SoC evolution,
the U-boot code is now a bit "confused" about which one to use!
Originally, the peripheral clock was used for both the TMU and the SCIF
(UART). However, today (March 2010), the ASC has largely replaced the
SCIF, and instead uses the "comms" clock.

The ASC driver was based on the SCIF one, and unfortunately the clocking
nomenclature was unchanged - hence the U-boot code for the SCIF and the
ASC both claim to use the "peripheral" clock, which is correct for the
former, and incorrect for the latter.

In summary, the clocking infrastructure needs to be overhauled and
brought up to date, as the peripheral clock often runs at a different
frequency to the comms clock.

This patch does *not* solve the general problem, but it provides a
short-term workaround for boards using the STx5197 SoC. It is envisaged
that a better fix that allows the comms and peripheral clocks to be
referenced correctly will be available shortly.

For further details, please see:
	https://bugzilla.stlinux.com/show_bug.cgi?id=8461

This patch has been tested on a 5197cab board by running the command
"sleep 30", and verifying the wall time was (by eye) 30 seconds.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>