[SH] Preparatory work for using new "poke interpreter".
This update logically does not change the behaviour of U-boot, it is
primarily a refactoring exercise to facilitate the ability to use one
of two possible "poke interpreters".

For many years, U-boot has used a hand-written "poke loop" and "poke
table", which performs the initial set of low-level pokes to setup
the PLLs, LMI, EMI, etc.  With the recent advent of the new romgen
tool, the set of "operations" understood by this "poke interpreter"
has evolved.  For a while, it was possible to map the new opcodes to
the old ones. However, with the latest versions of the targetpacks,
this is no longer possible.  Hence, it is now necessary for U-boot to
be able to use the latest version of the poke interpreter.

This update allows the user to choose either the "old" or the "new"
poke interpreter. However, as stated, this update provides no logical
change, and is just a refactoring operation, all the affected boards
have explicitly been configured to use the old poke interpreter.

/*
 * Only define the following if you want to use the "old"
 * poke loop table interpreter. This should *not* be
 * defined if you are now using the "romgen" tool to
 * generate contemporary poke tables. The latest generation
 * targetpacks may be incompatible with the old interpreter.
 */

This macro may be placed in the appropriate board's config.h file,
however, as use of this may be tightly coupled with the ".romgem" file,
the macro is placed in the init-<board>.h file instead.

To be clear, the absence of this macro implicitly will use the new
interpreter. This should be valid for all old poke tables, but to
reduce risk, this update will continue to use the old interpreter for
all pre-existing boards. Clearly, all new boards should exclusively use
the new interpreter.

	The old interpreter will uses the following files:
		cpu/sh/init_ram.S
		cpu/sh/old-pokeloop.S
		include/asm-sh/old-pokeloop.h

	The new interpreter will uses the following files:
		cpu/sh/init_ram.S
		cpu/sh/pokeloop.S
		include/asm-sh/pokeloop.h

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>