[SH] Optimization in poke interpreter (one load/cache line)
This is just an optimization of the code that copies the "poke table"
from RAM. Previously, the code explicitly loaded *every* word into the
operand cache. However, with the ST40 cache architecture, we only need
explicitly to load one element per cache line, and the other data items
in the same cache line will be automatically loaded as well. Hence,
the copy loop will execute about one-eighth as many iterations.

NOTE:	This does not result in an 8x speed up of the loop!
	The cache misses will still dominate the time!

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>