[SH] Added support for Ethernet for the 7141EUD Board
This update adds Ethernet support on the 7141EUD board, which uses
STMicroelectronics' STx7141 SoC.

This has been successfully tested with the on-chip GMAC #1, using the
on-board IC+ IP1001 PHY (U51). Tested on STi7141EUD-V1.0A-012.

Presently, I do not have access to a suitable H/W board to test with
the on-chip GMAC#0, using the off-board PHY (via J19). The code has been
written to use this GMAC+PHY combination, but in the absence of suitable
H/W, it has not been possible to test it.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>