[SH,SNAPSHOT] Add initial support for the MB837 (STx7108) board.
This update is a snapshot of the port for STMicroelectronics' MB837
MBoard, which uses the STx7108 SoC.

To date, the following are all known to work:
	RAM, 32-bit mode, RS-232, NOR, NAND and boot-from-NOR.

Further updates for Ethernet, SPI, etc, should be available soon.
All testing was been done with the MB837A mated with a MB705B peripheral
board.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>