[SH] Add compile-time test for SDRAM derivable configurations
There are several macros which define where SDRAM starts, how big it
is, and where various things in RAM are located.  Unfortunately, it is
possible to define these different macros such that the overall set is
mutually inconsistent!  In the future, it should be a goal to define only
TWO of these macros, and derive all the others automatically. To this
end, as a short-term fix, new code will issue compile-time diagnostics,
if the proposed derivable configurations would fail!

The two essential macros to be defined:
	CFG_SDRAM_BASE, CFG_SDRAM_SIZE

Derived Macros:
	CFG_LOAD_ADDR       = CFG_SDRAM_BASE
	CFG_MEMTEST_START   = CFG_SDRAM_BASE
	CFG_MEMTEST_END     = CFG_SDRAM_BASE + CFG_SDRAM_SIZE - 3MiB
	TEXT_BASE           = CFG_SDRAM_BASE + CFG_SDRAM_SIZE - 1MiB
	CFG_SE_SDRAM_WINDOW = CFG_SDRAM_SIZE - 1

This update will just issue a "#warning", if any of the derived macros
are not in agreement with the above.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>