[SH] Added boot-from-SPI support for MB796 (STx5289/STx5206)
This update includes support for boot-from-SPI on the MB796 board,
which uses either the STx5289 or STx5206 devices from STMicroelectronics.

This has been tested in both 29-bit and 32-bit modes, on a single MB796A
mated with a MB705B peripheral board.

Final testing was also done to confirm that all 3 FLASH boot capabilities
were working (NOR, NAND and SPI). Also, this testing was performed on
both cut 1.1 and cut 2.1 of the STx5289, on 32-bit mode. In addition
re-testing of USB and ethernet was performed. All testing was successful.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>