[SH] Add support for STMicroelectronics' HMP7100 board
This update adds support for STMicroelectronics HMP7100 board, which
was populated with a (cut 2.0) STx7109 SoC device.

This update includes tested support for: RS-232, NOR flash, Ethernet
(on-chip STMAC, off-chip SMSC LAN9118), and USB. In addition both 29-bit
and 32-bit (SE) modes were tested. Also, the board was successfully
booted from NOR flash in both 29-bit and 32-bit modes.

It should be noted that access to only a cut 2.0 chip was available and
tested, hence cut 3.0 is untested.

	The following commands were used to connect with GDB.

% sh4-linux-gdb -ex "sh4tp <STMC>:mb442_stx7109_cut20:st40,extclk=27,seuc=0"

% sh4-linux-gdb -ex "sh4tp <STMC>:mb442_stx7109_cut20:st40,extclk=27,seuc=1"

Please note the following:
-------------------------
The latests version of the mb442 TargetPack (which the HMP7100 uses)
differentiates between cut2.0 and cut 3.0 STx7109 silicon in SE-mode. One
of the key differences pertains as to whither the LMI registers and LMI
base should be moved.

	By default, we have the following:
		cut 2.0 has LMI Physical base at 0x04000000,
		cut 3.0 has LMI Physical base at 0x40000000.

Hence, we need to build U-Boot differently, depending on which cut of
silicon is on the board!

Only define the following macro if you are using a cut 2.0 STb7109 part,
and are *not* moving the LMI base to 0x40000000.

	#define CONFIG_SH_STB7109_CUT2

Addendum
--------
A HMP7100 board with STx7109 cut 3.2 was made available to me. I can now
confirm that this commit does indeed work (out-of-the-box) with cut 3.2
silicon of the STx7109 on the HMP7100 board. Hence, this update is now
known to work with both cut 2.0 and cut 3.2 versions of the STx7109.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>