Added instrumentation to measure byes/sec figures (for FLASH)
This update adds conditionally compiled code which will record and print
the elapsed time (and rate) around some of the key command-line utility
functions to read/write FLASH.  This is used to facilitate easy gathering
of raw bytes/sec figures for NAND, NOR, and SPI flash devices.

By default, this code is not enabled. To enable it, add the following
in an appropriate file (e.g. board configuration file).

	#define CONFIG_MEASURE_TIME

Note, the printing is done by the macro PRINT_MEASURED_TRANSFER_RATE,
this macro can be usurped, if a macro of that name is already defined
before "include/common.h". See "include/common.h" for more details.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>