[USB] Add fix to support "dodgy" USB devices (that need additional delays)
U-boot's USB stack can sometimes fail to recognise some USB memory sticks.
The problem, occurs because U-Boot assumes that USB devices enumerate
almost instantaneously, see usb_hub_configure(). This is fixed this
by adding some judicious delays.  Linux works because enumeration is
interrupt driven and this is not available in U-Boot.

For further details see:
	https://bugzilla.stlinux.com/show_bug.cgi?id=8941

The following macros can be used to enable/configure the amount of
additional delays that are added.

	CFG_USB_DODGY_DEVICES_SUPPORT
	CFG_USB_DODGY_DEVICES_INITIAL_DELAY
	CFG_USB_DODGY_DEVICES_LOOP_DELAY
	CFG_USB_DODGY_DEVICES_MAX_DELAY

Special thanks to Richard Marsh of Marsh Consulting Limited
for providing the basis of this fix in usb_hub_configure().

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>