[SH] Enable (by default) "bdinfo" to also dump EMI Bank information
This update changes the default behaviour of the "bdinfo" command for
SoCs based on the SH-4 CPU. For such boards, the "bdinfo" command
will (by default) also dump the EMI Bank information. Previously
this information was only enabled if the user explicitly defined
CONFIG_CMD_BDI_DUMP_EMI_BANKS.  This macro is now implicitly defined to be
"1".

To revert to the old behaviour the following may be used:

	#define CONFIG_CMD_BDI_DUMP_EMI_BANKS 0

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>