[SH] Allow 32MiB of NOR to be accessed on HMP7100 board.
On some HMP7100 boards, the NOR flash for CSAn (EMI Bank #0) has been
increased to 32MiB.  However, the default TargetPacks only provide a
region of 16MiB for all of CSAn.  The update will set:
	sizeof(EMI Bank #0) == 32 MiB,
hence allowing all of the NOR flash to be accessed.

It should be noted, that this will have the side-effect of logically
disabling CSBn, as EMI Bank #0 will now occlude EMI Bank #1.

See https://bugzilla.stlinux.com/show_bug.cgi?id=9065

This has been tested in both 29-bit and 32-bit (SE) modes, on a board with
only 8MiB, but it is fully expected to work with a board containing 32MiB.
In addition, the on-board SMSC MAC+PHY device still works correctly,
as it is on EMI Bank #2 (CSCn)

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>