[SH] Allow 256 MiB of LMISYS when booting linux on EUD7141 in SE-mode
For the EUD7141 board, this commit increased LMI #0 (LMISYS) UBA to allow
all 256 MiB of DDR to be seen when U-Boot boots linux in 32-bit mode.

Note: U-boot itself is still only configured to use 128MiB of LMISYS,
as this is is what the default TargetPacks support. However, when U-boot
boots linux, linux will be able to see all 256MiB of LMISYS.

For further details, see:
	https://bugzilla.stlinux.com/show_bug.cgi?id=9105

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>