[SH] Added USB (OHCI) support for STMicroelectronics' FLI7510 SoC
This commit adds USB (OHCI) support for STMicroelectronics' FLI7510 SoC,
using the FLDB (Freeman Lite Development Board), FLDB_GPD201.

Testing was on a single FLDB_GPD201 board, with a serial number of SZ015.

Technical Note:
--------------
In the OHCI controller, there is a register HcRhDescriptorA (+0x48),
which has a bit-field NPS (NoPowerSwitching). This bit-field is one (of
several) which controls the power switching modes for the root ports.
Normally, linux *enables* the NPS feature, but in contrast, U-Boot
normally *disables* the NPS feature.  The fact the linux and U-Boot
configure the NPS feature differently, is not normally a problem, as
most H/W systems appear to work in either configuration.

However, it now looks like the FLI7510/FLDB *is* "sensitive" to the
status of the NPS feature. (In contrast, all previous ST40 SoCs have not
exhibited such dependency.)  So, linux on the FLDB works because linux
usually enables the NPS feature, and U-Boot failed on the FLDB because
U-Boot usually disables the NPS feature.

Hence, the "fix" for the FLDB is to force U-Boot to enable the NPS feature.

Currently, is is unknown if this is due to something in the IP, in the
SoC, or on the board.

Signed-off-by: Sean McGoogan <Sean.McGoogan@st.com>